Make utility funcs and act on promises
